+++ src/SDKs/Azure.Base/data-plane/Azure.Base.Tests/Testing/Mocks.cs
@@ -106,6 +106,7 @@ namespace Azure.Base.Testing
             public override int Status => _status;
 
             public override Stream ResponseContentStream => throw new NotImplementedException();
+
             public override string CorrelationId { get; set; }
 
             public void SetStatus(int status) => _status = status;
